Move test script
